# server/docs/swagger.yaml
basePath: /api
definitions:
  appointment.AppointmentDetailPet:
    properties:
      avatar_url:
        type: string
      birth_date:
        type: string
      breed:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  appointment.AppointmentDetailResponse:
    properties:
      appointment_date:
        type: string
      doctor:
        $ref: '#/definitions/appointment.AppointmentDoctorSummary'
      id:
        type: string
      medical_record:
        $ref: '#/definitions/appointment.MedicalRecordResponse'
      owner_notes:
        type: string
      pet:
        $ref: '#/definitions/appointment.AppointmentDetailPet'
      previous_medical_history:
        type: string
      service_type:
        $ref: '#/definitions/models.ServiceType'
      status:
        $ref: '#/definitions/models.AppointmentState'
    type: object
  appointment.AppointmentDoctorSummary:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  appointment.AppointmentListItem:
    properties:
      appointment_date:
        type: string
      id:
        type: string
      pet:
        $ref: '#/definitions/appointment.AppointmentPetSummary'
      service_type:
        $ref: '#/definitions/models.ServiceType'
      status:
        $ref: '#/definitions/models.AppointmentState'
    type: object
  appointment.AppointmentOwnerSummary:
    properties:
      avatar_url:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  appointment.AppointmentPetSummary:
    properties:
      avatar_url:
        type: string
      breed:
        type: string
      name:
        type: string
    type: object
  appointment.CheckupDetails:
    properties:
      focus_area:
        type: string
      purpose:
        type: string
    required:
    - focus_area
    - purpose
    type: object
  appointment.CheckupMedicalDataDTO:
    properties:
      cleanliness_notes:
        type: string
      nutrition_recommendations:
        type: string
      palpation:
        type: string
      periodic_care_recommendations:
        type: string
    required:
    - cleanliness_notes
    - palpation
    type: object
  appointment.CreateAppointmentRequest:
    properties:
      appointment_date:
        type: string
      checkup:
        $ref: '#/definitions/appointment.CheckupDetails'
      owner_notes:
        type: string
      pet_id:
        type: integer
      previous_medical_history:
        type: string
      reminder_id:
        type: string
      service_type:
        allOf:
        - $ref: '#/definitions/models.ServiceType'
        enum:
        - vaccine
        - checkup
        - treatment
      treatment:
        $ref: '#/definitions/appointment.TreatmentDetails'
      vaccine:
        $ref: '#/definitions/appointment.VaccineDetails'
    required:
    - appointment_date
    - pet_id
    - service_type
    type: object
  appointment.CreateAppointmentResponse:
    properties:
      id:
        type: string
    type: object
  appointment.InternalAppointmentDetailResponse:
    properties:
      appointment_date:
        type: string
      doctor:
        $ref: '#/definitions/appointment.AppointmentDoctorSummary'
      id:
        type: string
      medical_record:
        $ref: '#/definitions/appointment.MedicalRecordResponse'
      owner:
        $ref: '#/definitions/appointment.AppointmentOwnerSummary'
      owner_notes:
        type: string
      pet:
        $ref: '#/definitions/appointment.AppointmentDetailPet'
      previous_medical_history:
        type: string
      service_type:
        $ref: '#/definitions/models.ServiceType'
      status:
        $ref: '#/definitions/models.AppointmentState'
    type: object
  appointment.MedicalRecordResponse:
    properties:
      checkup:
        $ref: '#/definitions/appointment.CheckupMedicalDataDTO'
      physical_examination:
        $ref: '#/definitions/appointment.PhysicalExaminationDTO'
      treatment:
        $ref: '#/definitions/appointment.TreatmentMedicalDataDTO'
      type:
        $ref: '#/definitions/models.ServiceType'
      vaccine:
        $ref: '#/definitions/appointment.VaccineMedicalDataDTO'
    type: object
  appointment.PhysicalExaminationDTO:
    properties:
      heart_rate:
        type: string
      physical_condition:
        type: string
      respiratory_rate:
        type: string
      temperature:
        type: number
      weight:
        type: number
    required:
    - physical_condition
    - temperature
    - weight
    type: object
  appointment.PrescriptionDTO:
    properties:
      dosage:
        type: string
      frequency:
        type: string
      name:
        type: string
    required:
    - dosage
    - frequency
    - name
    type: object
  appointment.RejectAppointmentRequest:
    properties:
      reason:
        type: string
    required:
    - reason
    type: object
  appointment.ReminderInputDTO:
    properties:
      description:
        type: string
      reminder_date:
        type: string
      service_type:
        allOf:
        - $ref: '#/definitions/models.ServiceType'
        enum:
        - vaccine
        - checkup
        - treatment
    required:
    - description
    - reminder_date
    - service_type
    type: object
  appointment.SaveMedicalRecordRequest:
    properties:
      checkup:
        $ref: '#/definitions/appointment.CheckupMedicalDataDTO'
      physical_examination:
        $ref: '#/definitions/appointment.PhysicalExaminationDTO'
      reminders:
        items:
          $ref: '#/definitions/appointment.ReminderInputDTO'
        type: array
      treatment:
        $ref: '#/definitions/appointment.TreatmentMedicalDataDTO'
      vaccine:
        $ref: '#/definitions/appointment.VaccineMedicalDataDTO'
    required:
    - physical_examination
    type: object
  appointment.SelectDoctorRequest:
    properties:
      doctor_id:
        type: integer
    required:
    - doctor_id
    type: object
  appointment.TreatmentDetails:
    properties:
      home_care_received:
        type: boolean
      observed_symptoms:
        items:
          type: string
        minItems: 1
        type: array
      symptom_duration:
        type: string
    required:
    - observed_symptoms
    - symptom_duration
    type: object
  appointment.TreatmentMedicalDataDTO:
    properties:
      clinical_symptoms:
        type: string
      diagnosis:
        type: string
      estimated_cost:
        minimum: 0
        type: number
      home_care_notes:
        type: string
      medical_actions:
        type: string
      prescriptions:
        items:
          $ref: '#/definitions/appointment.PrescriptionDTO'
        type: array
    required:
    - clinical_symptoms
    - diagnosis
    - estimated_cost
    - medical_actions
    - prescriptions
    type: object
  appointment.VaccineDetails:
    properties:
      vaccine_type:
        type: string
    required:
    - vaccine_type
    type: object
  appointment.VaccineMedicalDataDTO:
    properties:
      administration_date:
        type: string
      batch_number:
        type: string
      brand:
        type: string
      post_vaccine_reaction:
        type: string
      pre_vaccine_condition:
        type: string
      vaccine_type:
        type: string
    required:
    - administration_date
    - batch_number
    - brand
    - pre_vaccine_condition
    - vaccine_type
    type: object
...
  core.ActionResponse:
    properties:
      success:
        type: boolean
    type: object
...
  models.AppointmentState:
    enum:
    - Menunggu Konfirmasi
    - Diterima
    - Ditolak
    - Menunggu Dokter
    - Dalam Penanganan
    - Selesai
    - Selesai Administrasi
    type: string
    x-enum-varnames:
    - StateWaitingConfirmation
    - StateAccepted
    - StateRejected
    - StateWaitingDoctor
    - StateInTreatment
    - StateFinished
    - StateAdminFinished
...
  models.ServiceType:
    enum:
    - vaccine
    - checkup
    - treatment
    type: string
    x-enum-varnames:
    - ServiceTypeVaccine
    - ServiceTypeCheckup
    - ServiceTypeTreatment
...
  pet.PetAppointmentResponse:
    properties:
      appointment_date:
        type: string
      id:
        type: string
      service_type:
        $ref: '#/definitions/models.ServiceType'
      status:
        $ref: '#/definitions/models.AppointmentState'
    type: object
...
host: localhost:1323
info:
  contact: {}
  description: API for VetConnect Clinic Management System.
  title: VetConnect API
  version: "1.0"
paths:
  /appointments:
    get:
      description: Get a list of upcoming or past appointments for the current pet
        owner.
      parameters:
      - description: 'Filter: ''upcoming'' or ''past'''
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/appointment.AppointmentListItem'
            type: array
      security:
      - BearerAuth: []
      summary: Get Owner Appointments
      tags:
      - Appointment
    post:
      consumes:
      - application/json
      description: Create a new appointment request for a pet.
      parameters:
      - description: Create Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/appointment.CreateAppointmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/appointment.CreateAppointmentResponse'
      security:
      - BearerAuth: []
      summary: Create Appointment
      tags:
      - Appointment
  /appointments/{id}:
    get:
      description: Get details of a specific appointment for the authenticated owner.
      parameters:
      - description: Appointment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/appointment.AppointmentDetailResponse'
      security:
      - BearerAuth: []
      summary: Get Appointment Detail
      tags:
      - Appointment
...
  /internal/appointments:
    get:
      description: Get a list of appointments for staff, with optional status, date,
        and doctor filters.
      parameters:
      - description: Appointment Status
        in: query
        name: status
        type: string
      - description: Appointment Date (YYYY-MM-DD)
        in: query
        name: date
        type: string
      - description: Filter for doctor's own appointments
        in: query
        name: my_appointments
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/appointment.AppointmentListItem'
            type: array
      security:
      - BearerAuth: []
      summary: Get All Appointments (Internal)
      tags:
      - Appointment
  /internal/appointments/{id}:
    get:
      description: Get comprehensive appointment information for staff members.
      parameters:
      - description: Appointment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/appointment.InternalAppointmentDetailResponse'
      security:
      - BearerAuth: []
      summary: Get Internal Appointment Detail
      tags:
      - Appointment
  /internal/appointments/{id}/approve:
    post:
      description: Accepts a new appointment request, moving it to 'Accepted' state.
      parameters:
      - description: Appointment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.ActionResponse'
      security:
      - BearerAuth: []
      summary: Approve Appointment
      tags:
      - Appointment
  /internal/appointments/{id}/doctor-approve:
    post:
      description: Transition an appointment from 'Waiting Doctor' to 'In Treatment'
        state.
      parameters:
      - description: Appointment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.ActionResponse'
      security:
      - BearerAuth: []
      summary: Doctor Approve Appointment
      tags:
      - Appointment
  /internal/appointments/{id}/doctor-reject:
    post:
      consumes:
      - application/json
      description: Rejects an assigned appointment, moving it back to 'Accepted' state
        for reassignment.
      parameters:
      - description: Appointment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Rejection Reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/appointment.RejectAppointmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.ActionResponse'
      security:
      - BearerAuth: []
      summary: Doctor Reject Appointment
      tags:
      - Appointment
  /internal/appointments/{id}/medical-record:
    post:
      consumes:
      - application/json
      description: Save physical examination, service-specific data, and reminders
        for an appointment.
      parameters:
      - description: Appointment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Medical Record Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/appointment.SaveMedicalRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.ActionResponse'
      security:
      - BearerAuth: []
      summary: Save Medical Record
      tags:
      - Appointment
  /internal/appointments/{id}/reject:
    post:
      consumes:
      - application/json
      description: Rejects a new appointment request with a provided reason.
      parameters:
      - description: Appointment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Rejection Reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/appointment.RejectAppointmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.ActionResponse'
      security:
      - BearerAuth: []
      summary: Reject Appointment
      tags:
      - Appointment
  /internal/appointments/{id}/select-doctor:
    post:
      consumes:
      - application/json
      description: Assign a specific doctor to an appointment that has been accepted.
      parameters:
      - description: Appointment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Doctor ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/appointment.SelectDoctorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.ActionResponse'
      security:
      - BearerAuth: []
      summary: Select Doctor
      tags:
      - Appointment
...
  /pets/{id}/appointments:
    get:
      description: Get a list of all appointments for a specific pet owned by the
        current user.
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/pet.PetAppointmentResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get Pet Appointments
      tags:
      - Pet
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"